+++ lib/pages/Venda/venda.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:varejoMais/components/home_page_button.dart';
-import 'package:varejoMais/pages/Venda/vendaPage.dart';
+import 'package:varejoMais/pages/Venda/venda_page.dart';
 
-import 'package:varejoMais/platform_channel/platform_channel.dart';
+import 'package:varejoMais/shared/platform_channel/platform_channel.dart';
 
 class Venda extends StatelessWidget {
   Venda({super.key});
@@ -12,14 +11,9 @@ class Venda extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
-                child: Material(
-                            child: VendaPage(
-                            ),
-                ),
+      child: Material(
+        child: VendaPage(),
+      ),
     );
   }
 }
-
-
-
-
